move to rpc_invoker.py
