+++ src/amqp_invoker.py
@@ -1,12 +1,12 @@
 """Summary."""
-import asyncio
-import json
-from typing import Awaitable, Callable, Iterable
-from urllib.parse import urlparse
-
 import aio_pika
 import aiomisc
+import asyncio
+import json
+
 from retry import retry
+from typing import Awaitable, Callable, Iterable
+from urllib.parse import urlparse
 
 from src.function_invocable import FunctionInvocable
 from src.invoker import Invoker
@@ -74,7 +74,15 @@ class AmqpInvoker(Invoker):
         async with connection:
             async with connection.channel() as channel:
                 try:
-                    exchange = await channel.declare_exchange(name=self.exchange_name, type=aio_pika.ExchangeType.TOPIC, passive=False, durable=True, auto_delete=False, internal=False, arguments=None)
+                    exchange = await channel.declare_exchange(
+                        name=self.exchange_name,
+                        type=aio_pika.ExchangeType.TOPIC,
+                        passive=False,
+                        durable=True,
+                        auto_delete=False,
+                        internal=False,
+                        arguments=None
+                    )
                     queue = await channel.declare_queue(name=self.queue_name)
                     queue_error = await channel.declare_queue(name=f'{self.queue_name}_error')
 
@@ -97,7 +105,12 @@ class AmqpInvoker(Invoker):
 
         return connection
 
-    async def consume(self, channel: aio_pika.RobustChannel, queue: aio_pika.RobustQueue, callback: Callable[[aio_pika.IncomingMessage, aio_pika.RobustChannel], Awaitable[None]]) -> None:
+    async def consume(
+        self,
+        channel: aio_pika.RobustChannel,
+        queue: aio_pika.RobustQueue,
+        callback: Callable[[aio_pika.IncomingMessage, aio_pika.RobustChannel], Awaitable[None]]
+    ) -> None:
         """
         Binds a single consumer tag to `queue` and continuously consumes `aio_pika.IncomingMessage`s into `callback`.
 
@@ -129,16 +142,14 @@ class AmqpInvoker(Invoker):
                 # `self.do_work` is guaranteed to run on a separate thread, which is recommended for potentially long-running background tasks
                 async with self.do_work(data_in) as generator:
                     async for data_out in generator:
-                        await exchange.publish(message=aio_pika.Message(body=json.dumps(data_out).encode()),
-                                               routing_key=str(self._invocable.config.pubtopic))
+                        await exchange.publish(message=aio_pika.Message(body=json.dumps(data_out).encode()), routing_key=str(self._invocable.config.pubtopic))
 
             except Exception as err:  # pylint: disable=broad-except
                 data_in['error'] = str(err)
                 # TODO(ahuman-bean): cleaner error messages
-                await exchange.publish(message=aio_pika.Message(body=json.dumps(data_in).encode()),
-                                       routing_key=f'{self.queue_name}_error')
+                await exchange.publish(message=aio_pika.Message(body=json.dumps(data_in).encode()), routing_key=f'{self.queue_name}_error')
 
-    async def on_error(self, message: aio_pika.IncomingMessage, channel: aio_pika.RobustChannel) -> None:
+    async def on_error(message: aio_pika.IncomingMessage, channel: aio_pika.RobustChannel) -> None:
         """
         Stub for an error queue callback.
 
@@ -161,36 +172,8 @@ class AmqpInvoker(Invoker):
             payload: Lazily-evaluable wrapper around return values from `self._invocable.invoke` plus metadata
         """
         for data_out in self._invocable.invoke(data_in["data"]):
-            yield {"data": data_out, "key": str(self._invocable.config.pubtopic), "log": data_in.get("log", [])}
-
-
-class RPCInvoker(AmqpInvoker):
-    async def on_message(self, message: aio_pika.IncomingMessage, channel: aio_pika.RobustChannel) -> None:
-        """
-        Primary event queue callback.
-
-        Parameters:
-            message: Message object with convenience methods for acknowledgement
-            channel: Connection handle and state
-        """
-
-        # `message.process` will call `message.ack` upon `__aexit__` or `__exit__` (since no additional flags are passed to it),
-        # as well as handle requeueing and rejects if there are failures
-        async with message.process():
-            data_in: TYPE_PAYLOAD = dict(json.loads(message.body.decode('utf-8')))
-            try:
-                # `self.do_work` is guaranteed to run on a separate thread, which is recommended for potentially long-running background tasks
-                async with self.do_work(data_in) as generator:
-                    async for data_out in generator:
-                        out_message = aio_pika.Message(body=json.dumps(data_out).encode(),
-                                                       correlation_id=message.correlation_id)
-                        await channel.default_exchange.publish(
-                            message=out_message,
-                            routing_key=message.reply_to,
-                        )
-
-            except Exception as err:  # pylint: disable=broad-except
-                data_in['error'] = str(err)
-                # TODO(ahuman-bean): cleaner error messages
-                message = aio_pika.Message(body=json.dumps(data_in).encode())
-                await channel.default_exchange.publish(message=message, routing_key=message.reply_to)
+            yield {
+                "data": data_out,
+                "key": str(self._invocable.config.pubtopic),
+                "log": data_in.get("log", [])
+            }
